+++ src/basic_tokenizer.h
@@ -4,6 +4,6 @@
 #include "structs/table.h"
 #include "tokenizer_ops.h"
 
-void get_merges(GList ** ids, int n_merges, TokenPairToCountTable * token_merge_table, char * vocab[VOCAB_SIZE]);
+void get_merges(GList ** ids, long n_merges, TokenPairToCountTable * token_merge_table, char * vocab[VOCAB_SIZE]);
 
 #endif // BASIC_TOKENIZER_H!
