+++ src/tokenizer_ops.c
@@ -5,11 +5,11 @@
 
 void get_stats(GList * ids, TokenPairToCountTable * token_pair_counts) {
     for (GList * iterator = ids; iterator !=  NULL && iterator->next != NULL; iterator = iterator->next) { 
-        int *first_token = (int *) iterator->data;
-        int *second_token = (int *) iterator->next->data;
+        long *first_token = (long *) iterator->data;
+        long *second_token = (long *) iterator->next->data;
         TokenPair * pair = token_pair_new(*first_token, *second_token);
 
-        int value = table_lookup(token_pair_counts, pair);
+        long value = table_lookup(token_pair_counts, pair);
 
         if (value == -1) {
             table_insert_or_update(token_pair_counts, pair, 1);
@@ -19,11 +19,11 @@ void get_stats(GList * ids, TokenPairToCountTable * token_pair_counts) {
     }
 }
 
-void merge(GList * ids, GList ** new_pids, TokenPair pair, int id) {
+void merge(GList * ids, GList ** new_pids, TokenPair pair, long id) {
     for (GList * iterator = ids; iterator != NULL && iterator->next != NULL; iterator = iterator->next) {
-        int * new_id = malloc(sizeof(int));
-        int *first_token = (int *) iterator->data;
-        int *second_token = (int *) iterator->next->data;
+        long * new_id = malloc(sizeof(long));
+        long *first_token = (long *) iterator->data;
+        long *second_token = (long *) iterator->next->data;
         if (pair.first_token == *first_token && pair.second_token == *second_token) {
             iterator = iterator->next;
             *new_id = id;
@@ -35,10 +35,10 @@ void merge(GList * ids, GList ** new_pids, TokenPair pair, int id) {
     *new_pids = g_list_reverse(*new_pids);
 }
 
-void buffer_to_ids(unsigned char * buffer, GList ** pids, int sequence_length) { 
-    for (int i = 0; i < sequence_length; i++) {
-        int * id = malloc(sizeof(int));
-        *id = (int) buffer[i];
+void buffer_to_ids(unsigned char * buffer, GList ** pids, long sequence_length) { 
+    for (long i = 0; i < sequence_length; i++) {
+        long * id = malloc(sizeof(int));
+        *id = (long) buffer[i];
         *pids = g_list_prepend(*pids, id);
     }
     *pids = g_list_reverse(*pids);
@@ -46,7 +46,7 @@ void buffer_to_ids(unsigned char * buffer, GList ** pids, int sequence_length) {
 }
 
 void get_initial_vocab(char * vocab[VOCAB_SIZE]) {
-    for (int i = 0; i < 256; i++) {
+    for (long i = 0; i < 256; i++) {
         vocab[i] = (char *) malloc(2 * sizeof(char));
 
         if (vocab[i] == NULL) {
@@ -60,7 +60,7 @@ void get_initial_vocab(char * vocab[VOCAB_SIZE]) {
 }
 
 void vocab_free(char * vocab[VOCAB_SIZE]) {
-    for (int i = 0; i < VOCAB_SIZE; i++) {
+    for (long i = 0; i < VOCAB_SIZE; i++) {
         if (vocab[i] == NULL || vocab[i] == 0) {
             break;
         }
