+++ src/tokenizer_ops.h
@@ -6,11 +6,11 @@
 
 #define VOCAB_SIZE 267 
 
-void merge(GList * ids, GList ** new_pids, TokenPair pair, int id);
+void merge(GList * ids, GList ** new_pids, TokenPair pair, long id);
 
 void get_stats(GList * ids, TokenPairToCountTable * token_pair_counts);
 
-void buffer_to_ids(unsigned char * buffer, GList ** pids, int sequence_length);
+void buffer_to_ids(unsigned char * buffer, GList ** pids, long sequence_length);
 
 void get_initial_vocab(char * vocab[VOCAB_SIZE]);
